Refactor internal Map to match the new interface. Remove carto map integrator
