+++ src/index.js
@@ -22,7 +22,7 @@
 import { setDefaultAuth } from './setup/auth-service';
 import { setDefaultConfig } from './setup/config-service';
 import Viz from './Viz';
-import Map from './integrator/Map';
+import Map from './maps/Map';
 import Interactivity from './interactivity/Interactivity';
 import Layer from './layers/Layer';
 import * as expressions from './renderer/viz/expressions';
