+++ src/maps/Map.js
@@ -30,52 +30,49 @@ export default class Map {
         this._layers = new Set();
         this._hiddenLayers = new Set();
         this._repaint = true;
-        this.invalidateWebGLState = () => {};
         this._canvas = this._createCanvas();
         this._container.appendChild(this._canvas);
-        this._gl = this._canvas.getContext('webgl') || this._canvas.getContext('experimental-webgl');
+        this._gl = this._canvas.getContext('webgl');
         this._resizeCanvas(this._containerDimensions());
     }
 
-    addLayer (layer, beforeLayerID) {
-        layer.initialize();
+    on (event, callback) {
+        if (event === 'load') {
+            callback && callback();
+        }
+    }
 
+    addLayer (layer, beforeLayerID) {
         if (!this._layers.has(layer)) {
             this._layers.add(layer);
         }
 
-        window.requestAnimationFrame(this.update.bind(this));
+        layer.onAdd(this, this._gl);
+
+        window.requestAnimationFrame(this._update.bind(this));
     }
 
-    update (timestamp) {
-        // Don't re-render more than once per animation frame
-        if (this.lastFrame === timestamp) {
-            return;
+    removeLayer (layer) {
+        if (this._layers.has(layer)) {
+            this._layers.remove(layer);
         }
 
-        this.lastFrame = timestamp;
+        layer.onRemove(this);
+    }
 
-        this._drawBackground(this._background);
+    getZoom () {
+        return 0;
+    }
 
-        let loaded = true;
-        let animated = false;
+    getCenter () {
+        return { lat: 0, lng: 0 };
+    }
 
-        this._layers.forEach((layer) => {
-            const hasData = layer.hasDataframes();
-            const hasAnimation = layer.isAnimated();
-
-            if (hasData || hasAnimation) {
-                layer.$paintCallback();
-            }
-
-            loaded = loaded && hasData;
-            animated = animated || hasAnimation;
-        });
-
-        // Update until all layers are loaded or there is an animation
-        if (!loaded || animated) {
-            window.requestAnimationFrame(this.update.bind(this));
-        }
+    getBounds () {
+        return {
+            getNorthWest: () => ({ lat: 85, lng: -180 }),
+            getSouthWest: () => ({ lat: -85, lng: -180 })
+        };
     }
 
     changeVisibility (layer) {
@@ -100,6 +97,37 @@ export default class Map {
         if (this._hiddenLayers.has(layer)) {
             this._hiddenLayers.delete(layer);
             this._layers.add(layer);
+        }
+    }
+
+    _update (timestamp) {
+        // Don't re-render more than once per animation frame
+        if (this.lastFrame === timestamp) {
+            return;
+        }
+
+        this.lastFrame = timestamp;
+
+        this._drawBackground(this._background);
+
+        let loaded = true;
+        let animated = false;
+
+        this._layers.forEach((layer) => {
+            const hasData = layer.hasDataframes();
+            const hasAnimation = layer.isAnimated();
+
+            if (hasData || hasAnimation) {
+                layer.render();
+            }
+
+            loaded = loaded && hasData;
+            animated = animated || hasAnimation;
+        });
+
+        // Update until all layers are loaded or there is an animation
+        if (!loaded || animated) {
+            window.requestAnimationFrame(this._update.bind(this));
         }
     }
